chore: improve InputAddon exports
